Implement game sessions support
This PR adds support for the X-Tuffle-Game-ID header, which allows the
browser to persist the game state for some time on the server.

The game state will be restored accordingly using this header every time
the game starts up, which fixes a bug where the game state is gone when
the user refreshes even though the backend is still using that game.